refactor: dedup shared helpers across server and frontend
Extract repeated copy-paste helpers into shared modules:
- gitEnv() -> server/git-env.ts (was duplicated in 4 files)
- shellQuoteSingle() -> server/shell-quote.ts (3 files)
- taskWorkingDir() -> server/task-paths.ts (11 inline call sites)
- frontend timeAgo()/shortRepoName()/authorLabel() -> existing shared libs
