+++ server/chats.ts
@@ -10,6 +10,7 @@ import { getSettings } from './settings.js';
 import { getHarness } from './harnesses/index.js';
 import { hookBaseUrl } from './hook-base-url.js';
 import { childLogger } from './logger.js';
+import { shellQuoteSingle } from './shell-quote.js';
 import type { Agent } from './types.js';
 
 const execFile = promisify(execFileCb);
@@ -40,10 +41,6 @@ export interface CreateChatOptions {
   agent?: string | null;
   prompt?: string | null;
   harnessId?: string | null;
-}
-
-function shellQuoteSingle(s: string): string {
-  return `'${s.replace(/'/g, `'\\''`)}'`;
 }
 
 /**
