+++ server/diff-base.ts
@@ -1,6 +1,7 @@
 import { execFile as execFileCb } from 'child_process';
 import { promisify } from 'util';
 import { childLogger } from './logger.js';
+import { gitEnv } from './git-env.js';
 import type { Task } from './types.js';
 
 const execFile = promisify(execFileCb);
@@ -93,16 +94,6 @@ function cacheKey(cwd: string, baseBranch: string): string {
 
 function shortSha(sha: string): string {
   return sha.slice(0, 7);
-}
-
-// Strip GIT_* env vars so our git calls target the worktree we pass via -C,
-// not whatever repo an outer caller (e.g. a git hook) happens to be in.
-function gitEnv(): NodeJS.ProcessEnv {
-  const env: NodeJS.ProcessEnv = {};
-  for (const [k, v] of Object.entries(process.env)) {
-    if (!k.startsWith('GIT_')) env[k] = v;
-  }
-  return env;
 }
 
 /**
