+++ server/task-runner.ts
@@ -15,6 +15,7 @@ import { childLogger } from './logger.js';
 import type { RepoConfig } from './repo-config.js';
 import type { Task, Agent, UserTerminal, RunMode, Worktree } from './types.js';
 import { chatDirFor, chatSessionName } from './chats.js';
+import { shellQuoteSingle } from './shell-quote.js';
 import { computeMergeBase } from './git-commits.js';
 
 const logger = childLogger('task-runner');
@@ -78,10 +79,6 @@ async function getLastWindowIndex(session: string): Promise<number> {
 const PROMPT_FILE_CLEANUP_MS = 60000;
 
 const DISABLED_PLUGINS_IN_WORKTREES = ['remember@claude-plugins-official'] as const;
-
-function shellQuoteSingle(s: string): string {
-  return `'${s.replace(/'/g, `'\\''`)}'`;
-}
 
 function writeAgentLocalSettings(worktreePath: string): void {
   const claudeDir = path.join(worktreePath, '.claude');
