+++ src/components/PermissionPromptRow.tsx
@@ -1,15 +1,7 @@
 import { memo } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { timeAgo } from '@/lib/time';
 import type { PermissionPrompt } from '../../server/types';
-
-function timeAgo(dateStr: string): string {
-  const seconds = Math.floor((Date.now() - new Date(dateStr + 'Z').getTime()) / 1000);
-  if (seconds < 60) return 'just now';
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m ago`;
-  const hours = Math.floor(minutes / 60);
-  return `${hours}h ago`;
-}
 
 function abbreviateInput(toolInput: Record<string, unknown>): string {
   const command = toolInput.command || toolInput.file_path || toolInput.pattern || '';
